Re-enable compression feature
The DSP backend has now been improved to the point that the effect
works again. Applications that wish to enjoy dynamic range compression
must use the AudioEffect.ACTION_* intents to broadcast their session IDs
as specified by AudioEffect.

+++ src/com/bel/android/dspmanager/activity/DSPScreen.java
@@ -38,7 +38,7 @@ public final class DSPScreen extends PreferenceActivity {
 		super.onCreate(savedInstanceState);
 
 		try {
-			addPreferencesFromResource((Integer) this.getResources().getIdentifier(getSubPage() + "_preferences","xml",this.getPackageName()));
+			addPreferencesFromResource(this.getResources().getIdentifier(getSubPage() + "_preferences","xml",this.getPackageName()));
 		}
 		catch (Exception e) {
 			throw new RuntimeException(e);
